requirements.txt
beautifulsoup4
discord.py